コンポ－ネント分割とメモ化
