+++ src/components/atoms/RadioAtoms.tsx
@@ -1,13 +1,14 @@
 import { Radio } from "@chakra-ui/react";
-import { ReactNode, VFC } from "react";
+import { memo, VFC } from "react";
+import { RadioAtomsProps } from "../../types/PropsTypes";
 
-type Props = { children: ReactNode; value: string | number | undefined };
+type Props = RadioAtomsProps;
 
-export const RadioAtoms: VFC<Props> = (props: Props) => {
+export const RadioAtoms: VFC<Props> = memo((props: Props) => {
 	const { children, value } = props;
 	return (
 		<>
 			<Radio value={value}>{children}</Radio>
 		</>
 	);
-};
+});
