+++ src/components/atoms/SlideSwitch.tsx
@@ -1,8 +1,10 @@
 import { Switch } from "@chakra-ui/react";
 import { memo, VFC } from "react";
+import { SlideSwitchProps } from "../../types/PropsTypes";
 
-type Props = { onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void };
+type Props = SlideSwitchProps;
+
 export const SlideSwitch: VFC<Props> = memo((props: Props) => {
-	const { onChange } = props;
-	return <Switch onChange={onChange} size="md" />;
+	const { onChange, size, colorScheme, shadow } = props;
+	return <Switch onChange={onChange} size={size} colorScheme={colorScheme} shadow={shadow} />;
 });
