+++ src/components/molecule/RadioCategory.tsx
@@ -1,5 +1,5 @@
 import { RadioGroup, Stack } from "@chakra-ui/react";
-import { memo, useEffect, VFC } from "react";
+import { memo, VFC } from "react";
 import { useRecoilState } from "recoil";
 import { categoryState } from "../../globalState/categoryState";
 import { RadioAtoms } from "../atoms/RadioAtoms";
@@ -11,10 +11,6 @@ type Props = {
 export const RadioCategory: VFC<Props> = memo(() => {
 	const [checkedValue, setCheckedValue] = useRecoilState(categoryState);
 	const categoryList = ["メモ", "スケジュール", "TODOリスト"];
-
-	useEffect(() => {
-		console.log(checkedValue);
-	}, [checkedValue]);
 
 	return (
 		<RadioGroup onChange={setCheckedValue} value={checkedValue}>
