--- src/image/HeaderImage.tsx
@@ -1,4 +1,0 @@
-import { Image } from "@chakra-ui/react";
-export const HeaderImage = () => {
-	return <Image src={`${process.env.PUBLIC_URL}/business-pug-working-on-laptop.jpg`} fit="cover" />;
-};
