+++ src/router/Router.tsx
@@ -1,6 +1,6 @@
 import { ReactElement, VFC } from "react";
 import { Route, Routes } from "react-router-dom";
-import AppIntroduction from "../components/pages/AppIntroduction";
+import { AppIntroductionPage } from "../components/pages/AppIntroductionPage";
 import { Blank } from "../components/pages/Blank";
 import { Login } from "../components/pages/Login";
 import Page404 from "../components/pages/Page404";
@@ -9,7 +9,7 @@ import TopPage from "../components/pages/TopPage";
 export const Router: VFC = (): ReactElement => {
 	return (
 		<Routes>
-			<Route path="/" element={<AppIntroduction />}></Route>
+			<Route path="/" element={<AppIntroductionPage />}></Route>
 			<Route path="/login" element={<Login />} />
 			<Route path="/top" element={<TopPage />} />
 			<Route path="/blank" element={<Blank />} />
